bigdata/requirements.txt
fastapi~=0.114.0
uvicorn~=0.30.6
apscheduler
sqlalchemy~=2.0.35
databases~=0.9.0
aiomysql
pymysql
python-dotenv~=1.0.1
hdfs
pyspark
bs4~=0.0.2
pandas~=2.2.3
pydantic~=2.9.0
requests~=2.32.3
beautifulsoup4~=4.12.3